// src/reservation/enums/reservationStatus.enum.ts
export enum ReservationStatus {
  CANCELLED = 'CANCELLED',
  CLOSED = 'CLOSED',
  CONFIRMED = 'CONFIRMED',
  CONFIRMED_PREORDER = 'CONFIRMED_PREORDER',
  DISHONORED = 'DISHONORED',
  PENDING = 'PENDING',
  PENDING_PREORDER = 'PENDING_PREORDER',
  REJECTED = 'REJECTED',
  SERVE = 'SERVE',
  SERVE_PREORDER = 'SERVE_PREORDER',
}